requirements.txt
Pillow
tqdm
opencv-python
tensorflow-gpu
torchvision
